requirements.txt
streamlit
python-dotenv
groq
pandas
sentence-transformers
faiss-cpu
python-docx
PyPDF2
